Move bin behavior out of lib/ecstatic.js - this should help with non-standard envs and transpiling
